refactor: move getters to the Periodical type
