+++ src/periodic_config.rs
@@ -1,4 +1,4 @@
-use chrono::{DateTime, Days, Local, Months};
+use chrono::Local;
 use serde::Deserialize;
 
 use crate::prelude::*;
@@ -29,7 +29,7 @@ impl PeriodConfig {
     /// Attempts to get the configured file name associated with
     /// this Periodical.
     /// Returns a default format if not configured.
-    fn get_format(&self, period: Periodical) -> &str {
+    pub fn get_format(&self, period: Periodical) -> &str {
         self.fmt.as_deref().unwrap_or(match period {
             Periodical::Day => DEFAULT_DAY,
             Periodical::Week => DEFAULT_WEEK,
@@ -37,50 +37,10 @@ impl PeriodConfig {
             Periodical::Year => DEFAULT_YEAR,
         })
     }
-    /// Given a start date and an interval of Periodcals expressed as an uint,
-    /// will calculate the next interval date in time
-    /// with the correct formatting.
-    pub fn get_next(
-        &self,
-        date: DateTime<Local>,
-        interval: u32,
-        period: Periodical,
-    ) -> Option<String> {
-        let res = match period {
-            Periodical::Day => date.checked_add_days(Days::new(u64::from(interval))),
-            Periodical::Week => date.checked_add_days(Days::new(u64::from(interval * 7))),
-            Periodical::Month => date.checked_add_months(Months::new(interval)),
-            Periodical::Year => date.checked_add_months(Months::new(interval * 12)),
-        };
-
-        let fmt = self.get_format(period);
-        res.map(|f| f.format(fmt).to_string())
-    }
-    /// Given a start date and an interval of Periodcals expressed as an uint,
-    /// will calculate the next interval date in time
-    /// with the correct formatting.
-    pub fn get_prev(
-        &self,
-        date: DateTime<Local>,
-        interval: u32,
-        period: Periodical,
-    ) -> Option<String> {
-        let res = match period {
-            Periodical::Day => date.checked_sub_days(Days::new(u64::from(interval))),
-            Periodical::Week => date.checked_sub_days(Days::new(u64::from(interval * 7))),
-            Periodical::Month => date.checked_sub_months(Months::new(interval)),
-            Periodical::Year => date.checked_sub_months(Months::new(interval * 12)),
-        };
-
-        let fmt = self.get_format(period);
-        res.map(|f| f.format(fmt).to_string())
-    }
 }
 
 #[cfg(test)]
 mod tests {
-    use chrono::TimeZone;
-
     use super::*;
 
     #[test]
@@ -132,41 +92,5 @@ mod tests {
                 let got = config.format_file_name(period);
                 assert_eq!(want, got, "{desc}");
             });
-    }
-
-    #[test]
-    fn test_get_next() {
-        let date = Local.with_ymd_and_hms(2025, 12, 30, 0, 0, 0).unwrap();
-        let config = PeriodConfig::default();
-
-        let test_cases = [
-            (Periodical::Day, 1, "2025-12-31"),
-            (Periodical::Week, 1, "2026-W02"),
-            (Periodical::Month, 1, "2026-01"),
-            (Periodical::Year, 2, "2027"),
-        ];
-
-        test_cases.into_iter().for_each(|(period, interval, want)| {
-            let got = config.get_next(date, interval, period);
-            assert_eq!(Some(want.to_string()), got)
-        });
-    }
-
-    #[test]
-    fn test_get_prev() {
-        let date = Local.with_ymd_and_hms(2025, 12, 30, 0, 0, 0).unwrap();
-        let config = PeriodConfig::default();
-
-        let test_cases = [
-            (Periodical::Day, 1, "2025-12-29"),
-            (Periodical::Week, 1, "2025-W52"),
-            (Periodical::Month, 10, "2025-02"),
-            (Periodical::Year, 2, "2023"),
-        ];
-
-        test_cases.into_iter().for_each(|(period, interval, want)| {
-            let got = config.get_prev(date, interval, period);
-            assert_eq!(Some(want.to_string()), got)
-        });
     }
 }
